backend/requirements.txt
agno[pgvector]>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pandas>=2.0.0
openpyxl>=3.1.0
psycopg[binary]>=3.1.0
pgvector>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
